Fix issue with ASCII Characters and switch to OCLC
The Metadata decoder block was parsing incoming base64 records to
ASCII, which was throwing errors when it recieved UTF-8 text.
Updating the decoder (and catching any errors in the future) should
have resolved this issue.

This update also sets the manager to pass records to OCLC as the
first step